IMPALA-13472: Add the latest version of Breakpad
There is an issue with minidump stacks on ARM Redhat 8.
This adds a build of the most recent version of Breakpad
(e92bea3075) to see if it helps. This newer version also
includes direct support for compressed debuginfo (IMPALA-11907).

To make it easier to bump Breakpad versions, this switches
the build script to checkout Breakpad from git rather than
requiring an archive to be generated and uploaded to the
s3 bucket. This auto-detects the LSS version and checks it
out from git as well.

The patch that adds PPC64 support no longer applies cleanly
on the new version. Since there is no activity on PPC64 support
for Impala, this drops that patch for the new version of Breakpad.

Change-Id: I854e04c5ab68fc5aff7373c5c2197702dae01cb8