IMPALA-14759: Upgrade to boost 1.91.0
This bumps the version of Boost to 1.91.0. This includes a new
open addressing hash table and other updates.

Boost bumped its requirement for the C++ standard in some
libraries, and Avro was using an older C++ standard. This
bumps the version of Avro C++ to 1.12.2, which fixes that
issue. Avro has also been working to eliminate the Boost
dependency, and it is now only needed when building the tests.
This specifies AVRO_BUILD_TESTS=OFF to eliminate Avro C++'s
dependency on Boost

Testing:
 - Built Impala with the new Boost and ran tests

Change-Id: I669dd6d998a72ca855ec3a374bd31db17ce1c9b0
Reviewed-on: http://gerrit.cloudera.org:8080/24684
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>