IMPALA-11917: Upgrade to GCC 15 and LLVM 19
Upgrades to GCC 15, which also requires upgrading Clang to a newer
version. Upgrades to LLVM 19 as it's the first version that doesn't
show a performance regression in Impala; the update requires moving to
opaque pointers. Also builds LLVM with profile-guided optimization (PGO)
for better performance.

Updates binutils and breakpad for GCC 15 fixes. Patches others.

Updates Kudu to a newer version that links with mold to fix issues with
R_X86_64_GOTPCRELX and lld 11.

Removes some unused versions.

Change-Id: Ic7793748ff86d9c27530e135cb3b74319067a971